+++ srcs/ray.c
@@ -6,7 +6,7 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/22 17:51:22 by hthomas           #+#    #+#             */
-/*   Updated: 2020/03/03 16:27:35 by hthomas          ###   ########.fr       */
+/*   Updated: 2020/11/12 13:04:58 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ t_rgb		str_to_rgb(char *str)
 	return (char_to_rgb(tab[0], tab[1], tab[2]));
 }
 
-float	to_rad(float angle)
+double	to_rad(double angle)
 {
 	return (angle * M_PI * 0.5);
 }
@@ -56,9 +56,9 @@ t_ray		new_ray(const t_vect position, const t_vect direction)
 t_vect		set_vect_dir_cam(const t_camera *camera, const t_couple resolution, const int i, const int j)
 {
 	t_vect	vect_dir;
-	float	a;
-	float	b;
-	float	c;
+	double	a;
+	double	b;
+	double	c;
 	int		max;
 
 	a = j + 0.5 - (resolution.w) * 0.5;
